more generalized abvm.py
trying to turn this into an actual module. now has two functions, one of which will take a .h5 file generated from skynet and output a dictionary of mass number | abundance. the other will take an input of the list of dictionaries generated (if more than one dataset is being plotted) as well as a list of legend labels and create a plot. THINGS TO WORK ON: could benefit from a main function. legend might require order dictionaries = order of legend labels in order to correctly label. is that ok? dude idk how to fix that.